requirements.txt
accelerate==0.15.0
git+https://github.com/openai/CLIP.git
numpy
pandas
Pillow
PyYAML
requests
tqdm
urllib3
